# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Portal64 API
  description: |
    REST API for DWZ (Deutsche Wertungszahl) chess rating system.
    
    Provides access to player ratings, club information, and tournament data
    from the SVW (Schachverband Württemberg) chess federation databases.
    
    ## Player ID Format
    Player IDs follow the format: `{VKZ}-{PersonID}` (e.g., `C0101-1014`)
    where VKZ is the club identifier and PersonID is the unique person identifier.
    
    ## Club ID Format  
    Club IDs use the VKZ format: `{C}{NNNN}` (e.g., `C0101` for "Post-SV Ulm")
    
    ## Tournament ID Format
    Tournament IDs follow the format: `{Code}-{SubCode}-{Type}` where:
    - Code: Letter (A-Z) + year digit + week (e.g., B718=2017 week 18, C529=2025 week 29)  
    - Examples: `B718-A08-BEL`, `C413-612-DSV`, `C529-K00-HT1`
    
    ## Response Formats
    All endpoints support both JSON and CSV formats:
    - JSON: Default format or use `Accept: application/json` header
    - CSV: Use `Accept: text/csv` header or `?format=csv` query parameter
  version: 1.0.0
  contact:
    name: API Support
    email: support@svw.info
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.svw.info
    description: Production server
...
paths:
...
  /api/v1/tournaments:
    get:
      tags:
        - tournaments
      summary: Search tournaments
      description: Search tournaments by name, code, or other criteria
      parameters:
        - name: query
          in: query
          description: Search query (tournament name or code)
          schema:
            type: string
          example: "Hauptturnier"
        - name: limit
          in: query
          description: Maximum number of results (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 500
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [tname, finishedOn, rounds]
            default: finishedOn
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: filter_by
          in: query
          description: Filter by field
          schema:
            type: string
            enum: [year, type]
        - name: filter_value
          in: query
          description: Filter value
          schema:
            type: string
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: List of tournaments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentsResponse'
            text/csv:
              schema:
                type: string
  /api/v1/tournaments/{id}:
    get:
      tags:
        - tournaments
      summary: Get tournament by ID
      description: Get detailed information about a specific tournament
      parameters:
        - name: id
          in: path
          required: true
          description: Tournament ID/Code in format C529-K00-HT1
          schema:
            type: string
          example: "C529-K00-HT1"
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Tournament details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentResponse'
            text/csv:
              schema:
                type: string
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tournaments/recent:
    get:
      tags:
        - tournaments
      summary: Get recent tournaments
      description: Get recently finished tournaments
      parameters:
        - name: days
          in: query
          description: Number of days to look back
          schema:
            type: integer
            minimum: 1
            default: 30
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 500
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: List of recent tournaments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TournamentResponse'
            text/csv:
              schema:
                type: string
  /api/v1/tournaments/date-range:
    get:
      tags:
        - tournaments
      summary: Get tournaments by date range
      description: Get tournaments within a specific date range
      parameters:
        - name: start_date
          in: query
          required: true
          description: Start date in YYYY-MM-DD format
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: end_date
          in: query
          required: true
          description: End date in YYYY-MM-DD format
          schema:
            type: string
            format: date
          example: "2024-12-31"
        - name: query
          in: query
          description: Search query
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 500
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: List of tournaments in date range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentsResponse'
            text/csv:
              schema:
                type: string
        '400':
          description: Bad request (invalid date format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    TournamentResponse:
      type: object
      properties:
        id:
          type: string
          description: Tournament ID/Code
          example: "C529-K00-HT1"
        name:
          type: string
          description: Tournament name
          example: "SUI - SEM 2025 - Hauptturnier 1 Disentis"
        code:
          type: string
          description: Tournament code
          example: "C529-K00-HT1"
        type:
          type: string
          description: Tournament type
          example: "Open"
        organization:
          type: string
          description: Organizing body
          example: "Swiss Chess Federation"
        rounds:
          type: integer
          description: Number of rounds
          example: 9
        start_date:
          type: string
          format: date-time
          description: Tournament start date
          example: "2025-01-15T10:00:00Z"
        end_date:
          type: string
          format: date-time
          description: Tournament end date
          example: "2025-01-22T18:00:00Z"
        status:
          type: string
          enum: [open, registration_open, running, finished, cancelled]
          description: Tournament status
          example: "finished"
        participant_count:
          type: integer
          description: Number of participants
          example: 128
...
    TournamentsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/TournamentResponse'
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of results
          example: 150
        limit:
          type: integer
          description: Limit per page
          example: 20
        offset:
          type: integer
          description: Current offset
          example: 0
        count:
          type: integer
          description: Number of results in current page
          example: 20
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "Resource not found"
        message:
          type: string
          description: Detailed error message
          example: "The requested player could not be found"